Don't merge sequential unstyled tags
This change fixes #738 by ensuring that `convertFromHTMLToContentBlocks` counts unstyled blocks as blocks and by moving the logic to suppress a block's type into that function.

Previously the existence of any tag other than the unstyled tag would cause unstyled blocks not to be parsed as blocks (but, curiously, the aliases of the unstyled tag would be parsed as blocks). This is mean to handle a case like

```
<div><h2>Hello</h2></div>
```

ensuring that the final block is a `header-two` but breaks with examples like

```
<h2>Hello</h2>
<div>world!</div>
<div>Goodbye!</div>
```

which parses as two blocks with the two divs being combined into a single unstyled block.

The core issue is that not counting the unstyled tag when an interesting tag appears anywhere in the document is a global problem to a local issue: the fix is to instead consider each tree under a block individually with unstyled blocks deferring their final type to whatever the next lowest block type is.